Initial Frankenscript
Still needs changes to make all consistent, particularly with the data which could use some more specific naming, and I want to add automatic saving of the plots to the results folder, imports and loading are somewhat organised. Should run just fine, fixed some errors which are still prevalent in the non-franken files.